WIP: More restructuring. Added ability to indicate module name in @defcomp by using a name with a dot separating module and component name, e.g., "Mimi.adder".
