+++ src/Mimi.jl
@@ -1,25 +1,22 @@
+__precompile__(false)
+
 module Mimi
 
 using DataStructures
 using DataFrames
 using Distributions
 using NamedArrays
-using MacroTools
 
 export
     @defcomp,
-    @deftimestep,
     ComponentDef,
     ComponentKey,
-    ConnectorCompMatrix,        # deprecated
-    ConnectorCompVector,        # deprecated
     MarginalModel,
     Model,
     Timestep,
     TimestepMatrix,
     TimestepVector,
     addcomponent,
-    adder,
     adddimension,
     addparameter,
     components,
@@ -59,17 +56,14 @@ import
 include("core/metainfo.jl")
 include("modelinstance/mi_types.jl")
 include("modelinstance/clock.jl")
-include("modelinstance/deftimestep_macro.jl")
 include("modelinstance/run.jl")
 
 include("core/mimi_types.jl")
 include("core/timestep_arrays.jl")
 include("core/references.jl")
-include("core/defcomp2.jl")
+include("core/defcomp.jl")
 include("core/build.jl")
 include("core/mimi-core.jl")
-
-include("helpercomponents/marginalmodel.jl")
 
 include("utils/graph.jl")
 include("utils/plotting.jl")
@@ -78,8 +72,8 @@ include("utils/lint_helper.jl")
 
 # Components are defined here to allow pre-compilation to work
 function __init__()
-    include("helpercomponents/adder.jl")
-    include("helpercomponents/connectorcomp.jl")   
+    compdir = joinpath(dirname(@__FILE__), "helpercomponents")
+    load_comps(compdir)
 end
 
 end # module
