+++ src/core/defcomp.jl
@@ -1,192 +1,203 @@
 #
-# This file is deprecated, but left here for reference for now.
+# @defcomp and supporting functions
 #
+using MacroTools
 
-function curr_module_name()
-    return Base.module_name(current_module())
-end
+Debug = false
 
-# text for constructing module name and component types from @defcomp name
-const module_prefix = "_mimi_module_"
-const comp_suffix = "Impl"
-
-#
-# TO DO: 
-# - Change this to generate a struct with all the relevant data rather than with
-#   expressions. Generate the expressions from the struct. Incorporate improvements
-#   from defcomp_new using @capture.
-# - Add the code from deftimestep_macro.jl to incorporate the timestep function into @defcomp
-#
-"""
-    @defcomp name begin expressions... end
-
-Define a new component.
-"""
-macro defcomp(name, ex)
-    module_name = curr_module_name()
-
-    resetvarsdef = Expr(:block) # don't accumulate expressions here
-    metavardef = Expr(:block)
-    metapardef = Expr(:block)
-    metadimdef = Expr(:block)
-
-    numarrayparams = 0
-
-    for line in ex.args
-        if line.head==:(=) && line.args[2].head==:call && line.args[2].args[1]==:Index
-            dimensionName = line.args[1]
-
-            push!(metadimdef.args, :(adddimension(curr_module_name(), $(Expr(:quote,name)), $(QuoteNode(dimensionName)) )))
-        elseif line.head==:(=) && line.args[2].head==:call && line.args[2].args[1]==:Parameter
-            if isa(line.args[1], Symbol)
-                parameterName = line.args[1]
-                parameterType = :Number
-            elseif line.args[1].head==:(::)
-                parameterName = line.args[1].args[1]
-                parameterType = line.args[1].args[2]
-            else
-                error()
-            end
-
-            kws = collectkw(line.args[2].args)
-
-            # Get description and unit, if provided
-            description = get(kws, :description, "")
-            unit = get(kws, :unit, "")
-
-            if haskey(kws, :index)
-                parameterIndex = kws[:index].args
-
-                if length(parameterIndex) <= 2 && parameterIndex[1] == :time
-                    numarrayparams += 1
-                end
-
-                pardims = Array{Any}(0)
-                for l in parameterIndex
-                    push!(pardims, l)
-                end
-
-                push!(metapardef.args, :(addparameter($(QuoteNode(module_name)), $(Expr(:quote,name)), $(QuoteNode(parameterName)), $(esc(parameterType)), $(pardims), $(description), $(unit))))
-            else
-                push!(metapardef.args, :(addparameter($(QuoteNode(module_name)), $(Expr(:quote,name)), $(QuoteNode(parameterName)), $(esc(parameterType)), [], $(description), $(unit))))
-            end
-
-        elseif line.head == :(=) && line.args[2].head == :call && line.args[2].args[1 ]== :Variable
-            if isa(line.args[1], Symbol)
-                variableName = line.args[1]
-                variableType = :Number
-            elseif line.args[1].head == :(::)
-                variableName = line.args[1].args[1]
-                variableType = line.args[1].args[2]
-            else
-                error()
-            end
-
-            kws = collectkw(line.args[2].args)
-
-            # Get description and unit, if provided
-            description = get(kws, :description, "")
-            unit = get(kws, :unit, "")
-
-            if haskey(kws, :index)
-                variableIndex = kws[:index].args
-
-                vardims = Array{Any}(0)
-                for l in variableIndex
-                    push!(vardims, l)
-                end
-
-                push!(metavardef.args, :(addvariable(curr_module_name(), $(Expr(:quote,name)), $(QuoteNode(variableName)), $(esc(variableType)), $(vardims), $(description), $(unit))))
-
-                if variableType == :Number
-                    push!(resetvarsdef.args,:($(esc(Symbol("fill!")))(s.Variables.$(variableName),$(esc(Symbol("NaN"))))))
-                end
-            else
-                push!(metavardef.args, :(addvariable(curr_module_name(), $(Expr(:quote,name)), $(QuoteNode(variableName)), $(esc(variableType)), [], $(description), $(unit))))
-
-                if variableType == :Number
-                    push!(resetvarsdef.args,:(s.Variables.$(variableName) = $(esc(Symbol("NaN")))))
-                end
-            end
-        elseif line.head == :line
-            # TBD: use MacroTools.rmlines() to get rid of these
-        else
-            error("Unknown expression.")
-        end
+function debug(msg)
+    if Debug
+        println(msg)
     end
-
-    module_name = Symbol(string(module_prefix, name))
-    module_def = :(eval(current_module(), :(module $module_name end)))
-
-    call_expr = Expr(:call,
-                    Expr(:curly,
-                        Expr(:., 
-                            Expr(:., 
-                                Expr(:., 
-                                    :Main, 
-                                    QuoteNode(Symbol(current_module()))), 
-                                QuoteNode(module_name)),
-                            QuoteNode(Symbol(string(name, comp_suffix)))),
-                        :T, :OFFSET, :DURATION, :FINAL),
-                    :indices)
-
-    callsignature = Expr(:call, Expr(:curly, Symbol(name), :T, :OFFSET, :DURATION, :FINAL), :(::Type{T}), :(::Type{Val{OFFSET}}),:(::Type{Val{DURATION}}),:(::Type{Val{FINAL}}))
-    for i in 1:numarrayparams
-        offset   = Symbol("OFFSET$i")
-        duration = Symbol("DURATION$i")
-
-        push!(call_expr.args[1].args, offset)
-        push!(call_expr.args[1].args, duration)
-
-        push!(callsignature.args[1].args, offset)
-        push!(callsignature.args[1].args, duration)
-        push!(callsignature.args, :(::Type{Val{$(QuoteNode(offset))}}))
-        push!(callsignature.args, :(::Type{Val{$(QuoteNode(duration))}}))
-
-    end
-    push!(callsignature.args, :indices)
-    # println(call_expr)
-    # println(callsignature)
-    # println(Expr(:function, callsignature, call_expr))
-
-    return gen_component_expr(name, module_def, resetvarsdef, metavardef, metapardef, metadimdef, callsignature, call_expr)
 end
 
 #
-# TODO: this needs to be largely rewritten. Should take a struct with all the values rather than all these args.
+# Main> MacroTools.prewalk(replace_dots, :(p.foo[1,2] = v.bar))
+# :((get_property(p, Val(:foo)))[1, 2] = get_property(v, Val(:bar)))
 #
-function gen_component_expr(name, module_def, resetvarsdef, metavardef, metapardef, metadimdef, callsignature, call_expr)
-    expr = quote
+# Main> MacroTools.prewalk(replace_dots, :(p.foo = v.bar[1]))
+# :(set_property!(p, Val(:foo), (get_property(v, Val(:bar)))[1]))
+#
+function _replace_dots(ex)
+    debug("\nreplace_dots($ex)\n")
 
-        abstract type $name <: Mimi.ComponentState end
+    if @capture(ex, obj_.field_ = rhs_)
+        return :(set_property!($obj, Val($(QuoteNode(field))), $rhs))
+    
+    elseif @capture(ex, obj_.field_)
+        return :(get_property($obj, Val($(QuoteNode(field)))))
 
-        import Mimi.run_timestep
-        import Mimi.init
-        import Mimi.resetvariables
+    elseif @capture(ex, obj_.field_[args__] = rhs_)
+        return :(get_property($obj, Val($(QuoteNode(field))))[$(args...)] = $rhs)
 
-        function resetvariables(s::$name) 
-            $resetvarsdef
-        end
+    elseif @capture(ex, obj_.field_[args__])
+        return :(get_property($obj, Val($(QuoteNode(field))))[$(args...)])
 
-        # ComponentKey supplies current module name when only the component name is given
-        key = ComponentKey($(QuoteNode(name)))
-
-        # Create an empty component
-        addcomponent(key)
-
-        $metavardef
-        $metapardef
-        $metadimdef
-        $module_def
-
-        eval($(esc(Symbol(string(module_prefix, name)))), generate_comp_expressions(key))
-
-        # RP: not sure what DA was attempting to do here:
-        # callsignature.args[1].args[1] = $esc(Symbol(name)) # DA: how to do this?
-
-        function $name{$(callsignature.args[1].args[2:end]...)}($(callsignature.args[2:end]...))
-            $call_expr
-        end
+    else
+        #debug("No dots to replace")
+        return ex
     end
+end
+
+function _generate_run_func(comp_name, args, body)
+    module_name = Base.module_name(current_module())
+
+    # wrap list of body expressions into a single expr for prewalk()
+    block = Expr(:block)
+    block.args = body
+    block = MacroTools.prewalk(_replace_dots, block)
+    body = block.args
+
+    func = :(
+        function run_timestep(::Val{$(QuoteNode(module_name))}, ::Val{$(QuoteNode(comp_name))}, $(args...))
+            $(body...)
+        end
+    )
+    debug("func: $func")
+    return func
+end
+
+
+function _check_for_known_argname(name)
+    if !(name in (:description, :unit, :index))
+        error("Unknown argument name: '$name'")
+    end
+end
+
+function _check_for_known_element(name)
+    if !(name in (:Variable, :Parameter, :Index))
+        error("Unknown component element type: '$name'")
+    end
+end
+
+# Generates an expression to construct a Variable or Parameter
+function _generate_var_or_param(elt_type, name, datatype, dimensions, desc, unit)
+    func_name = elt_type == :Parameter ? :addparameter : :addvariable
+    expr = :($func_name(comp, $(QuoteNode(name)), $datatype, $dimensions, $desc, $unit))
+    debug("Returning: $expr\n")
     return expr
+end
+
+function _generate_index_expr(name, args, vartype)
+    debug("  Index $name")
+
+    # Args are not permitted; we attempt capture only to check syntax
+    if length(args) > 0
+        error("Index $name: arguments to Index() are not permitted")
+    end
+
+    # Ditto types for Index, e.g., region::Foo = Index()
+    if vartype != nothing
+        error("Index $name: Type specification ($vartype) is not supported")
+    end
+
+    expr = :(adddimension(comp, $(QuoteNode(name))))
+    return expr
+end
+
+_generate_index_expr(name::Symbol) = _generate_index_expr(name, [], nothing)
+
+#
+# Parse a @defcomp definition, converting it into a series of function calls that
+# create the corresponding ComponentDef instance. At model build time, the ModelDef
+# (including its ComponentDefs) will be converted to a runnable model.
+#
+macro defcomp(comp_name, ex)
+    known_dims = Set{Symbol}()
+    
+    @capture(ex, elements__)
+    debug("Component $comp_name")
+
+    # Allow explicit definition of module to define component in
+    if @capture(comp_name, modname_.cmpname_)       # e.g., Mimi.adder
+        comp_name = cmpname
+    else
+        modname = Base.module_name(current_module())
+    end
+    
+    # We'll return a block of expressions that will define the component.
+    # First, we add the empty component and assign it to `comp`.
+    result = quote comp = addcomponent($(QuoteNode(modname)), $(QuoteNode(comp_name))) end
+
+    function addexpr(expr)
+        push!(result.args, expr)
+    end
+
+    for elt in elements
+        debug("elt: $elt")
+
+        if @capture(elt, function run(args__) body__ end)
+            # Save the expression that will store the run_timestep function definition, and
+            # translate dot notation to get/setproperty. The func is created at build time.
+            expr = _generate_run_func(comp_name, args, body)
+            run_expr = :(set_run_expr(comp, $(QuoteNode(expr))))
+            addexpr(run_expr)
+            continue
+        end
+
+        if ! @capture(elt, (name_::vartype_ | name_) = elt_type_(args__))
+            error("Element syntax error: $elt")
+        end
+
+        # vartype = vartype == nothing ? ::Float64 : vartype
+        # debug("name: $name, vartype: $vartype, elt_type: $elt_type, args: $args")
+
+        # elt_type is one of {:Variable, :Parameter, :Index}
+        if elt_type == :Index
+            expr = _generate_index_expr(name, args, vartype)
+            push!(known_dims, name)
+            addexpr(expr)
+
+        elseif elt_type in (:Variable, :Parameter)
+            debug("  $elt_type $name")
+            desc = unit = ""
+            dimensions = Array{Symbol}(0)
+
+            for arg in args
+                debug("    arg: $arg")
+                if @capture(arg, argname_ = value_)
+                    _check_for_known_argname(argname)
+
+                elseif @capture(arg, argname_)
+                    # If it's an unknown arg, report *that* rather than the missing value
+                    _check_for_known_argname(argname)
+                    error("Argument '$arg' of $elt_type $name is missing a value")
+
+                else
+                    error("Badly formatted argument: $arg")
+                end
+
+                if @capture(arg, description = value_)
+                    desc = value
+
+                elseif @capture(arg, unit = value_)
+                    unit = value
+
+                elseif @capture(arg, index = [dims__])
+                    debug("    dims: $dims")
+                    append!(dimensions, dims)
+
+                    # Add undeclared dimensions on-the-fly
+                    for dim in dims
+                        if ! (dim in known_dims)
+                            addexpr(_generate_index_expr(dim))
+                            push!(known_dims, dim)
+                        end
+                    end
+                end
+            end
+
+            dims = Tuple(dimensions) # just for printing
+            debug("    index $dims, unit '$unit', desc '$desc'")
+
+            datatype = vartype == nothing ? Number : vartype
+            expr = _generate_var_or_param(elt_type, name, datatype, dimensions, desc, unit)
+            push!(result.args, expr)
+
+        else
+            error("Unrecognized element type: $elt_type")
+        end
+    end
+
+    return rmlines(result)
 end
