+++ src/helpercomponents/connectorcomp.jl
@@ -6,15 +6,12 @@ using Mimi
 #   required machinery using the functional API rather than using @defcomp.
 #
 
-@defcomp ConnectorCompVector begin
+@defcomp Mimi.ConnectorCompVector begin
     input1 = Parameter(index = [time])
     input2 = Parameter(index = [time])
-    output = Variable(index = [time])
+    output =  Variable(index = [time])
 
-    function run(s::ConnectorCompVector, ts::Timestep)
-        p = s.Parameters
-        v = s.Variables
-
+    function run(p, v, d, ts)
         if hasvalue(p.input1, ts)
             v.output[ts] = p.input1[ts]
         elseif hasvalue(p.input2, ts)
@@ -25,22 +22,18 @@ using Mimi
     end
 end
 
-@defcomp ConnectorCompMatrix begin
+@defcomp Mimi.ConnectorCompMatrix begin
     regions = Index()
 
     input1 = Parameter(index = [time, regions])
     input2 = Parameter(index = [time, regions])
-    output = Variable(index = [time, regions])
+    output =  Variable(index = [time, regions])
 
-    function run(s::ConnectorCompMatrix, ts::Timestep)
-        p = s.Parameters
-        v = s.Variables
-        d = s.Dimensions
-
+    function run(p, v, d, ts)
         for r in d.regions
             if hasvalue(p.input1, ts, r)
                 v.output[ts, r] = p.input1[ts, r]
-            elseif hasvalue(p.input2, ts, r)
+            elseif hasvalue(p.input2, t, r)
                 v.output[ts, r] = p.input2[ts, r]
             else
                 error("Neither of the inputs to ConnectorComp have data for the current timestep: $(gettime(ts)).")
